sometimes ember addons need to specify in what order they should load
the short version is: this works

the long version is:
I'm not sure.

The offending line in `wnyc-web-client` is calling 
`this.get('session').syncBrowserId()`.

the `session` service is initially defined in `ember-simple-auth` and 
extended in `nypr-auth` to add the `syncBrowserId` method.

Since `nypr-auth` extends the `session` service provided by 
`ember-simple-auth`, it specifies (using a similar feature) that it must 
come after `ember-simple-auth` in the load order, as well as 
`nypr-publisher-lib` and `nypr-audio-services`, which both also depend 
on `ember-simple-auth`.

That's where my trail runs cold. Setting this value in this repo seems 
to work :facepalm: